check authorization prior to action on backend; 117

+++ mean-course/backend/routes/posts.js
@@ -56,7 +56,8 @@ router.post(
 );
 
 router.put(
-  "/:id",
+  "/:id", 
+  checkAuth,
   multer({ storage: storage }).single("image"),
   (req, res, next) => {
     let imagePath = req.body.imagePath;
@@ -70,9 +71,12 @@ router.put(
       content: req.body.content,
       imagePath: imagePath
     });
-    console.log(post);
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-      res.status(200).json({ message: "Update successful!" });
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+      if(result.nModified > 0){
+        res.status(200).json({ message: "Update successful!" });
+      }else{
+        res.status(401).json({ message: "Not Authorized!" });
+      }
     });
   }
 );
@@ -109,10 +113,14 @@ router.get("/:id", (req, res, next) => {
   });
 });
 
-router.delete("/:id",checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
+router.delete("/:id", checkAuth, (req, res, next) => {
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
     console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
+    if (result.n > 0) {
+      res.status(200).json({ message: "Delete successful!" });
+    } else {
+      res.status(401).json({ message: "Not Authorized!" });
+    }
   });
 });
